group5/requirements.txt
# Django
Django==3.0.6
djangorestframework==3.11.0

# Django extensions
django-extensions==2.2.9
django-extra-fields==2.0.5

# Linting
flake8==3.8.1
black==19.10b0

# Misc
Pillow==7.1.2
requests==2.23.0